requirements.txt
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
matplotlib==3.0.2
netCDF4==1.4.2
notebook==5.7.4
numpy==1.15.4
pandas==0.24.1
pytest==4.3.0
qiskit-aer==0.1.1
qiskit-aqua==0.4.2
qiskit-terra==0.8.0
seaborn==0.9.0
xarray==0.11.3
xyzpy==0.3.1
